+++ clients/admin-ui/src/features/system/DescribeSystemStep.tsx
@@ -2,7 +2,7 @@ import { Box, Button, Heading, Stack, useToast } from "@fidesui/react";
 import { SerializedError } from "@reduxjs/toolkit";
 import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query/fetchBaseQuery";
 import { Form, Formik } from "formik";
-import React, { useState } from "react";
+import React, { useMemo, useState } from "react";
 import * as Yup from "yup";
 
 import {
@@ -16,6 +16,7 @@ import {
   defaultInitialValues,
   FormValues,
   transformFormValuesToSystem,
+  transformSystemToFormValues,
 } from "~/features/system/form";
 import {
   useCreateSystemMutation,
@@ -29,21 +30,39 @@ const ValidationSchema = Yup.object().shape({
   system_type: Yup.string().required().label("System type"),
 });
 
+const SystemHeading = ({ system }: { system?: System }) => {
+  const isManual = !system;
+  const headingName = isManual
+    ? "your new system"
+    : system.name ?? "this system";
+
+  return (
+    <Heading as="h3" size="lg">
+      Describe {headingName}
+    </Heading>
+  );
+};
+
 interface Props {
   onCancel: () => void;
   onSuccess: (system: System) => void;
   abridged?: boolean;
-  initialValues?: FormValues;
+  system?: System;
 }
 
 const DescribeSystemStep = ({
   onCancel,
   onSuccess,
   abridged,
-  initialValues: passedInInitialValues,
+  system: passedInSystem,
 }: Props) => {
-  const isEditing = !!passedInInitialValues;
-  const initialValues = passedInInitialValues ?? defaultInitialValues;
+  const initialValues = useMemo(
+    () =>
+      passedInSystem
+        ? transformSystemToFormValues(passedInSystem)
+        : defaultInitialValues,
+    [passedInSystem]
+  );
   const [createSystem] = useCreateSystemMutation();
   const [updateSystem] = useUpdateSystemMutation();
   const [isLoading, setIsLoading] = useState(false);
@@ -51,6 +70,14 @@ const DescribeSystemStep = ({
   const systemOptions = systems
     ? systems.map((s) => ({ label: s.name ?? s.fides_key, value: s.fides_key }))
     : [];
+  const isEditing = useMemo(
+    () =>
+      Boolean(
+        passedInSystem &&
+          systems?.some((s) => s.fides_key === passedInSystem?.fides_key)
+      ),
+    [passedInSystem, systems]
+  );
 
   const toast = useToast();
 
@@ -100,10 +127,8 @@ const DescribeSystemStep = ({
       {({ dirty, values }) => (
         <Form>
           <Stack spacing={10}>
-            <Heading as="h3" size="lg">
-              {/* TODO FUTURE: Path when describing system from infra scanning */}
-              Describe your system
-            </Heading>
+            <SystemHeading system={passedInSystem} />
+
             <div>
               By providing a small amount of additional context for each system
               we can make reporting and understanding our tech stack much easier
@@ -121,8 +146,8 @@ const DescribeSystemStep = ({
                 id="fides_key"
                 name="fides_key"
                 label="System key"
-                // TODO FUTURE: This tooltip text is misleading since at the moment for MVP we are manually creating a fides key for this resource
-                tooltip="System keys are automatically generated from the resource id and system name to provide a unique key for identifying systems in the registry."
+                disabled={isEditing}
+                tooltip="System keys are automatically generated from the resource id and system name to provide a unique key for identifying systems in the registry. Manually changing this may make it hard to keep track of your systems."
               />
               <CustomTextInput
                 id="description"
@@ -174,8 +199,11 @@ const DescribeSystemStep = ({
                 type="submit"
                 variant="primary"
                 size="sm"
-                // if isEditing, always allow going to the next step
-                disabled={isEditing ? false : !dirty}
+                isDisabled={
+                  isLoading ||
+                  // If this system was created from scratch, the fields must be edited.
+                  (!passedInSystem && !dirty)
+                }
                 isLoading={isLoading}
                 data-testid="confirm-btn"
               >