+++ scripts/setup/dataset.py
@@ -2,10 +2,11 @@ import logging
 from typing import Dict
 
 import requests
-import setup.constants as constants
 import yaml
 
 from fides.api.ops.api.v1 import urn_registry as urls
+
+from . import constants
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)