+++ tests/ops/integration_tests/saas/test_sentry_task.py
@@ -8,11 +8,19 @@ import requests
 from fides.api.ops.graph.graph import DatasetGraph
 from fides.api.ops.models.privacy_request import PrivacyRequest
 from fides.api.ops.schemas.redis_cache import Identity
+from fides.api.ops.service.connectors import get_connector
 from fides.api.ops.task import graph_task
 from fides.api.ops.task.filter_results import filter_data_categories
 from fides.api.ops.task.graph_task import get_cached_data_for_erasures
 from tests.ops.graph.graph_test_util import assert_rows_match
 from tests.ops.test_helpers.saas_test_utils import poll_for_existence
+
+
+@pytest.mark.skip(reason="Pending account resolution")
+@pytest.mark.integration_saas
+@pytest.mark.integration_sentry
+def test_sentry_connection_test(sentry_connection_config) -> None:
+    get_connector(sentry_connection_config).test_connection()
 
 
 @pytest.mark.skip(reason="Pending account resolution")
@@ -29,7 +37,7 @@ async def test_sentry_access_request_task(
     """Full access request based on the Sentry SaaS config"""
 
     privacy_request = PrivacyRequest(
-        id=f"test_saas_access_request_task_{random.randint(0, 1000)}"
+        id=f"test_sentry_access_request_task_{random.randint(0, 1000)}"
     )
     identity = Identity(**{"email": sentry_identity_email})
     privacy_request.cache_identity(identity)
@@ -270,13 +278,17 @@ def sentry_erasure_test_prep(sentry_connection_config, db):
 async def test_sentry_erasure_request_task(
     db, policy, sentry_connection_config, sentry_dataset_config
 ) -> None:
-    """Full erasure request based on the Sentry SaaS config. Also verifies issue data in access request"""
+    """
+    Full erasure request based on the Sentry SaaS config.
+    Also verifies issue data in access request.
+    """
+
     erasure_email, issue_url, headers = sentry_erasure_test_prep(
         sentry_connection_config, db
     )
 
     privacy_request = PrivacyRequest(
-        id=f"test_saas_access_request_task_{random.randint(0, 1000)}"
+        id=f"test_sentry_erasure_request_task_{random.randint(0, 1000)}"
     )
     identity = Identity(**{"email": erasure_email})
     privacy_request.cache_identity(identity)